backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.4.0
gunicorn==22.0.0
qiskit==2.2.3
qiskit-aer==0.17.2
qiskit-ibm-provider==0.7.1
pytest==8.2.2
pytest-django==4.8.0